Refactoring to split grpc, api related and specific versions in different files
