+++ src/kessel/inventory/v1/index.ts
@@ -1,0 +1,10 @@
+import { KesselInventoryHealthServiceClient } from "./health";
+import {
+  ClientBuilderFactory
+} from "../../grpc";
+
+
+
+export const ClientBuilder = ClientBuilderFactory(
+  KesselInventoryHealthServiceClient,
+);
